+++ src/ApiManagement/ApiManagement/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed a bug in `Get-AzApiManagementTenantGitAccess` cmdlet.
 
 ## Version 2.3.0
 * Added new `Sync-AzApiManagementKeyVaultSecret` cmdlet.